[Infer] Fix 11 Infer errors/warnings

+++ AsyncDisplayKit/Details/NSIndexSet+ASHelpers.m
@@ -68,9 +68,9 @@
   NSMutableString *result = [NSMutableString stringWithString:@"{ "];
   [self enumerateRangesUsingBlock:^(NSRange range, BOOL * _Nonnull stop) {
     if (range.length == 1) {
-      [result appendFormat:@"%zu ", range.location];
+      [result appendFormat:@"%zu ", (unsigned long)range.location];
     } else {
-      [result appendFormat:@"%zu-%zu ", range.location, NSMaxRange(range) - 1];
+      [result appendFormat:@"%zu-%u ", (unsigned long)range.location, NSMaxRange(range) - 1];
     }
   }];
   [result appendString:@"}"];
